refactor missing references
