+++ src/components/FeedbackForms/MissingReferences/index.ts
@@ -1,0 +1,1 @@
+export * from './ReferencesTable';
