+++ src/components/FeedbackForms/index.ts
@@ -1,1 +1,2 @@
 export * from './MissingRecord';
+export * from './MissingReferences';
